requirements.txt
# Dependencies
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.117.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.13.1
vertexai==0.0.1
Pillow==10.1.0
google-auth-oauthlib==1.2.0
requests==2.31.0


# Optional dependencies for local development
ipython==8.17.2
pytest==7.4.3


# videoplayer 

fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.3.0
aiofiles==23.2.1